add support for descriptions and multiple numbers per item
