--- constants/Communities.ts
@@ -1,45 +1,0 @@
-export const Communities = [{
-  btn: {
-    display: 'BIPOC Community',
-  },
-  params: {
-    section: 'bipoc',
-    title: 'BIPOC Support Hotlines',
-  },
-},
-{
-  btn: {
-    display: 'LGBTQA+ Community',
-  },
-  params: {
-    section: 'lgbtqa',
-    title: 'LGBTQA+ Support Hotlines',
-  },
-},
-{
-  btn: {
-    display: 'Immigrant Community',
-  },
-  params: {
-    section: 'immigrant',
-    title: 'Immigrant Support Hotlines',
-  },
-},
-{
-  btn: {
-    display: 'Youth Community',
-  },
-  params: {
-    section: 'homeless',
-    title: 'Homeless Support Hotlines',
-  },
-},
-{
-  btn: {
-    display: 'Elderly Community',
-  },
-  params: {
-    section: 'elderly',
-    title: 'Elderly Support Hotlines',
-  },
-}];
