+++ constants/index.ts
@@ -1,3 +1,2 @@
-export { Communities } from './Communities';
 export { Layout } from './Layout';
 export { Styles } from './Styles';
