+++ screens/index.ts
@@ -1,0 +1,6 @@
+export { AboutScreen } from './AboutScreen';
+export { CategoryScreen } from './CategoryScreen';
+export { CitySelectScreen } from './CitySelectScreen';
+export { NotFoundScreen } from './NotFoundScreen';
+export { PhoneNumberListScreen } from './PhoneNumberListScreen';
+export { SettingsScreen } from './SettingsScreen';
