+++ types.tsx
@@ -1,7 +1,7 @@
 import { RouteProp } from '@react-navigation/native';
 import { StackNavigationProp } from '@react-navigation/stack';
 
-export type tCacheUnion = 'Boston' | 'Chicago' | 'New York' | 'San Francisco Bay Area';
+export type tCacheUnion = 'Boston' | 'Chicago' | 'New York' | 'Los Angeles' | 'San Francisco Bay Area';
 
 /**
  * @description global types file. should just live in a declarations folder, ideally
